Initial Structure
ModelOrderReduction using Data-Driven methods often requires some data
reduction techniques to begin with. As a result, it is sensible to
abstract this out.

PCA which is used in Galerkin based POD is one
linear method, that Chris already refers to in Issue 2.

Non-linear data reduction methods like the Diffussion Maps is also
something that is in the pipeline.

Ofcourse, VAEs are well known to represent a state vector in latent
space, so this is a no-brainer, as there are some interesting model
reduction methods, that could benefit from this.